Update CTL for interface
Can now specify service type to emulate for testing clients. This
code goes nowhere yet though but at least the workflow is now
visible.

Next step is updating the TestNG classes and capability tests to
not use an IUT document, then to start an HTTP server for the
test session.