Replace 'iut' parameter with 'servicetype'
Modifies tests and classes to remove IUT and TEST_SUBJECT as well
as temp file creation.

As this is a client testing suite, we are not opening a connection
to a remote implementation to test. Instead we are running a
simple OGC Web Service implementation and inspecting a client's
connection.

The 'servicetype' parameter determines what type of OGC Web
Service we are emulating for the client.

At the moment it is a String parameter, perhaps an enum of allowed
values would be more strict?