Add dummy code for HTTP server singleton
This singleton will refer to the HTTP server instance that will
run in a separate thread from the test suite. This server will
then start with the test run parameters (if it is not already
running), and then when a test session is started a new path is
registered for the Secure Client to access.

When the Secure Client connects to the URL, the HTTP Server will
pass the details of that connection to the conformance test suites
for validation.

After the test session ends, the test session path is un-registered
from the HTTP Server.

If the Secure Client fails to connect to the HTTP Server within a
timeout value, a timeout failure should be triggered.

When the test suite is ran under an IDE or as an all-in-one JAR
file, the singleton nature of the HTTP Server will not have any
notable effect over a single-use initialization of an HTTP server
for a single test run.

In the case of running under TEAM Engine, the singleton is needed
for multiple TEAM Engine test sessions to be ran simultaneously
under different paths, otherwise the different test sessions would
conflict when binding the HTTP server to the same host and port.