Start embedded TestServer on test run
An embedded Jetty server starts on port 10080 when the first test
session is started. It runs in a separate thread. When a test
session is ran, the test controller will wait for the test server
to receive a request from the secure client at "localhost:10080/test"
or will error out if the timeout of 120 seconds is reached.

Once the request is received, the test server will stop blocking the
test session thread and the actual conformance classes will run.
The plan is to pass the request details from the test server into
the conformance classes as the "Implementation Under Test". (I am
not sure if only a single request needs to be passed in, or will
it be an array of requests that need to be analyzed?)

Currently most of this is hard-coded and only a single test session
can be ran at a time as only a single servlet is registered. I have
tested this with running the AIO JAR and hitting /test in a web
browser and it finishes the test without any exceptions.