Fix javadoc configuration for `site` goal
Without this sub-configuration for the `javadoc` plugin under the
`site` plugin, it would default to using an outdated URL for the
TestNG javadocs; an outdated URL causes a fatal error. The fatal
error then would cause any `mvn clean install` to fail.

This fix forces a working URL for the TestNG javadocs.