Add note about Content element
For Partial Capabilities, the Content element may be represented
as <Layer> in WMS, <Contents> in WMTS/WCS, <FeatureTypeList> in
WFS, and potentially other named elements for other services.

See section 8.2.2 of the specification for details.

Support for Partial Capabilities falls under Use Case III,
Protected Service / Private Data / Public Catalogue, which is not
in the scope of Testbed 14. It should be added to this ETS
eventually.