Deploy HTTPS on the TestServer Jetty instance
Connecting to the test server via plain HTTP no longer works,
HTTPS scheme must be used. Example:

$ curl -i -k
https://127.0.0.1:10080/asdfasdf?request=GetCapabilities&service=WMS

`-k` must be used as the certificate is self-signed. Certificate
is generated in the user home directory using:

$ cd $HOME
$ keytool -keystore keystore -storepass "ets-security-client" \
  -genkey -alias dummy-key \
  -dname "cn=Test User, ou=None, o=None, c=us"

This will be replaced with a robust way to specify the keystore path
and keystore password.