Serialize client requests to temp file
Serialization preserves the secure client requests after the
servlet request has been closed. Saving the request metadata to a
file allows me to pass that file path as a String to the
conformance class test methods.

I did not see a way to pass this data directly in the test run
properties.