Create sub-project for shaded libraries
A sub project uses the Maven Shade plugin to rename the package
namespace for the javax.servlet API, so that it will have a
different name and not conflict with the javax.servlet API loaded
by Tomcat (or similar) when run under TEAM Engine.

Jetty is also shaded, as the references to javax.servlet need to
be rewritten in it too so that it uses the shaded servlet API and
doesn't try to load the servlet API that Tomcat might have loaded.

This causes the following packages to be renamed, when doing
imports for classes:

javax.servlet.*     => org.opengis.cite.servlet.*
org.eclipse.jetty.* => org.opengis.cite.jetty.*

The sub-project also locks the Jetty server version and Servlet
API versions. Any update for those libraries must first be
applied there, then the sub-project version must be bumped and
then incremented in the main project POM.xml.