Omit JKS details from test results
These details may be semi-sensitive on a production system where
the JKS details should only be known by the system administrator
and not by any test users.

This test removes them from being sent to the Test Suite, and
removes the validation step before the Test Suite. The validation
step can be removed as it is ran before the Test Server was
started.