Handle exceptions for TestNG
Even if exceptions occur during test preparation, the TestNG
executor is still called with the exception as a test run
property. This means test users can see what exceptions were
raised that caused the test to fail, and they won't need to read
the server log to do it.

I added a class that wraps Properties objects and Documents, to
replace HashMap usage. This class allows me to send a test run
property document with a single entry for the exception.