Change documentation for authentication
It had been possible to specify multiple authentication methods,
but I have changed my mind. It would have been more difficult to
determine in the test server which authentication method the test
user really wanted to use, and to anticipate the correct number
of request/responses. Instead it will be easier to isolate to a
single authentication method only.