Remove saml-idp sample script
The node.js saml-idp does not seem to support CORS, which is
necessary for an AJAX secure client (used for testing). saml-idp
also does not have an HTTP Basic authentication option.

I intend to replace it with an external tool that is more hackable
and is based on Ruby instead.