Remove incomplete web test client
While it would have been easy to use, AJAX has some major issues
when it is redirected to the Identity Provider. The SP redirect
destroys the "Origin" header (part of CORS), so the IdP blocks the
redirected request from the client. The client has NO WAY of
detecting the redirect nor stopping the redirect, as the
XMLHttpRequest standard blocks this.