Use AssertionConsumerServiceURL for SAML 2
AssertionConsumerServiceURL tells the IdP where the response
should be sent. The secure client must then retrieve that URL
from the response in the "SubjectConfirmationData" element's
attribute "Recipient".

The URL is also available in the "Response" element's attribute
for "Destination", although I don't think the client is supposed
to use that.